pyproject.toml
[tool.poetry]
name = "health-sync"
version = "0.1.0"
description = "Automate collecting daily health metrics from multiple sources and append normalized rows into a Google Sheet."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "health_sync" }]

[tool.poetry.dependencies]
python = "^3.11"
typer = {version = "^0.12.5", extras=["all"]}
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
python-dotenv = "^1.0.1"
structlog = "^24.2.0"
httpx = "^0.27.2"
tenacity = "^9.0.0"
# Google APIs
google-api-python-client = "^2.147.0"
google-auth = "^2.35.0"
google-auth-httplib2 = "^0.2.0"

# Web server for Apple Health webhook
fastapi = "^0.115.2"
uvicorn = {version = "^0.32.0", extras=["standard"]}

# Playwright for scraping
playwright = "^1.48.0"

# Timezone handling
pytz = "^2024.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
vcrpy = "^6.0.1"
pytest-asyncio = "^0.24.0"
ruff = "^0.6.8"
black = "^24.8.0"
isort = "^5.13.2"
mypy = "^1.11.2"
pre-commit = "^4.0.1"

[tool.poetry.scripts]
health-sync = "health_sync.cli:app"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "C4", "DTZ", "TID", "SIM", "PL"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
